+++ pastry/src/rice/p2p/glacier/v2/messaging/GlacierMessage.java
@@ -53,7 +53,7 @@ public abstract class GlacierMessage implements RawMessage {
    *
    * @return This message's priority
    */
-  public byte getPriority() {
+  public int getPriority() {
     return LOW_PRIORITY;
   }
