+++ pastry/src/rice/pastry/messaging/PJavaSerializedDeserializer.java
@@ -24,9 +24,9 @@ public abstract class PJavaSerializedDeserializer extends JavaSerializedDeserial
     super(pn);
   }
 
-  public abstract Message deserialize(InputBuffer buf, short type, byte priority, NodeHandle sender) throws IOException;
+  public abstract Message deserialize(InputBuffer buf, short type, int priority, NodeHandle sender) throws IOException;
   
-  public rice.p2p.commonapi.Message deserialize(InputBuffer buf, short type, byte priority, rice.p2p.commonapi.NodeHandle sender) throws IOException {
+  public rice.p2p.commonapi.Message deserialize(InputBuffer buf, short type, int priority, rice.p2p.commonapi.NodeHandle sender) throws IOException {
     rice.p2p.commonapi.Message ret = deserialize(buf, type, priority, (NodeHandle)sender);
     if (ret == null) return super.deserialize(buf, type, priority, sender);
     return ret;
