+++ pastry/src/rice/tutorial/lesson3/MyMsg.java
@@ -36,7 +36,7 @@ public class MyMsg implements Message {
   /**
    * Use low priority to prevent interference with overlay maintenance traffic.
    */
-  public byte getPriority() {
+  public int getPriority() {
     return Message.LOW_PRIORITY;
   }
 }
